MP-302 swarm 出票分页列表

+++ miner/miner-server/src/main/java/com/mei/hui/miner/entity/SwarmTicket.java
@@ -18,9 +18,9 @@ public class SwarmTicket {
 
     private Integer nodePort;
 
-    private Byte type;
+    private Integer type;
 
-    private Byte state;
+    private Integer state;
 
     private BigDecimal parValue;
